move root files to root folder
