+++ src/ensembl/src/root/rootReducer.ts
@@ -1,9 +1,9 @@
 import { combineReducers } from 'redux';
 import { connectRouter } from 'connected-react-router';
 
-import global from './global/globalReducer';
-import header from './header/headerReducer';
-import browser from './content/app/browser/browserReducer';
+import global from '../global/globalReducer';
+import header from '../header/headerReducer';
+import browser from '../content/app/browser/browserReducer';
 
 const rootReducer = (history: any) =>
   combineReducers({
